+++ src/controllers/UserRouter.js
@@ -1,9 +1,10 @@
 const express = require("express");
-const { userModel } = require("../models/UserModel");
+const { userModel, postModel } = require("../models/models");
+
 const userRouter = express.Router();
 
 userRouter.get("/", async (req, res) => {
-    let result = await userModel.find({});
+    let result = await userModel.find({}).populate("posts");
     
     console.log(result)
     res.json({
@@ -13,7 +14,7 @@ userRouter.get("/", async (req, res) => {
 });
 
 userRouter.get("/:id", async (req, res) => {
-    let result = await userModel.findById(req.params.id);
+    let result = await userModel.findById(req.params.id).populate("posts");
 
     res.json({
         message: "Viewing profile by id",
