+++ src/schemas/TradeHistorySchema.js
@@ -4,12 +4,10 @@ const tradeHistorySchema = new mongoose.Schema({
     // have to link the user who the trade was made with
     tradesHistory:{
         // replaced with Mongoose Obj ID from User
-        type: { type: mongoose.Schema.Types.ObjectId, ref: 'Post_Model' }
+        // type: { type: mongoose.Schema.Types.ObjectId, ref: 'Post_Model' }
+        type: [String]
     },
 });
-
-// const tradeHistoryModel = mongoose.model('TradeHistory_Model', tradeHistorySchema);
-// module.exports = tradeHistoryModel;
 
 
 module.exports = {
