+++ src/utils/seed copy.js
@@ -1,10 +1,9 @@
-// import default users and posts to
-// start with some interesting fake posts created by fake users
+// import default users to start with some interesting fake users
 const defaultUsersData = require('./defaultUsers.json')
 const defaultPostsData = require('./defaultPosts.json')
 
-// Import modes and schemas from models.js
-const { userModel, postModel } = require("../models/models.js")
+const { userModel } = require("../models/UserModel.js")
+const { postModel } = require('../models/PostModel.js');
 
 
 // import database connect
@@ -13,8 +12,7 @@ const { dbConnect, dbDisconnect, dbClear } = require("../database");
 async function seedUsers() {
     let seed_defaultUsers = await userModel.insertMany(defaultUsersData);
 
-    console.log("seed_defaultUsers");
-    // console.log(seed_defaultUsers);
+    console.log(seed_defaultUsers);
     return seed_defaultUsers;
 }
 
@@ -35,8 +33,7 @@ async function seedPosts(userData) {
 
     let seed_defaultPosts = await postModel.insertMany(defaultPostsData_map);
 
-    console.log("seed_defaultPosts");
-    // console.log(seed_defaultPosts);
+    console.log(seed_defaultPosts);
     return seed_defaultPosts;
 }
 
@@ -48,13 +45,11 @@ async function seed(){
 
     let newUsers = await seedUsers();
     let newPosts = await seedPosts(newUsers);
-
-    // let newUsersWithModelsData = await seedUsers(newPosts);
-
+    let newUsersWithModelsData = await seedUsers(newPosts);
     // let defaulUsers = await seedUsers();
     // let defaulPosts = await seedPosts(defaulUsers);
 
-    console.log("All Data seeded");
+    console.log("Data seeded");
     // disconnect database
     dbDisconnect();
 }
